Updated context management logic
